nlp/requirements.txt
pydantic==1.10.8
datasets==2.12.0
evaluate==0.4.0
mlflow==2.3.2
accelerate==0.20.3
nltk==3.8.1
torch==2.0.1
transformers==4.29.2
pandas==2.0.2
modin==0.22.1
wordcloud==1.9.2
matplotlib==3.7.1
optimum[neural-compressor]==1.8.7
neural-compressor==2.1.1
webvtt-py==0.4.6
intel_extension_for_pytorch==2.0.100
scikit-learn==1.2.2
scikit-learn-intelex==2023.1.1
